Update README.md
Analysis for sales to know busiest days of week, times of day when most orders are placed to target marketing and type of products. The department which has the highest frequency of product orders, brand loyalty, customer loyalty, customer demographic and orders based on region, type of foods, family status and ordering habits based on the profiles.